backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
apyori==1.1.2
blinker==1.9.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
greenlet==3.3.1
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
nltk==3.9.2
numpy==2.4.2
packaging==26.0
pillow==12.1.1
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dotenv==1.2.1
regex==2026.1.15
requests==2.32.5
rich==14.3.2
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.5
wrapt==2.1.1
pandas==3.0.0
